kern/abi: decruft desc structs
The TaskDesc/RegionDesc and friends were originally generated by the
nascent Hubris build system as a series of 32-bit words concatenated
together, so it was important for them to be (1) predictable in size and
(2) repr(C). This is no longer the case, so these properties are no
longer mandatory.

In addition to simplifying the code and removing distracting cruft,
removing former reserved word padding and changing the sizes of some
types to be consistent throughout shrinks the task tables by several
bytes of flash per task.

+++ sys/kern/src/task.rs
@@ -59,7 +59,7 @@ impl Task {
         region_table: &'static [&'static RegionDesc],
     ) -> Self {
         Task {
-            priority: abi::Priority(descriptor.priority as u8),
+            priority: abi::Priority(descriptor.priority),
             state: if descriptor.flags.contains(TaskFlags::START_AT_BOOT) {
                 TaskState::Healthy(SchedState::Runnable)
             } else {
